Added UserInputUrl class
Moved main and URL-handling methods to new class.
